refactor(ProductViewController): view 패키지 생성, 이동
