+++ src/main/java/gift/view/ProductViewController.java
@@ -1,5 +1,6 @@
-package gift.controller;
+package gift.view;
 
+import gift.controller.ProductController;
 import gift.model.Product;
 import java.util.List;
 import org.springframework.stereotype.Controller;
